refactoring the issues-table component
