+++ src/utils.js
@@ -1,3 +1,4 @@
+//fetch the issues for react project from github
 export async function fetchIssues(page, perPage) {
   try {
     const response = await fetch(
@@ -14,10 +15,16 @@ export async function fetchIssues(page, perPage) {
   }
 }
 
+//Get the apropriate text color for the background color received
 export const textColorForBackground = backgroundColor => {
+  //get red, green and blue integer values of the color
   const r = parseInt(backgroundColor.slice(0, 2), 16);
   const g = parseInt(backgroundColor.slice(2, 4), 16);
   const b = parseInt(backgroundColor.slice(4, 6), 16);
+
+  //calculate the lume
   const lume = (r * 299 + g * 587 + b * 114) / 1000;
+
+  //return the apropriate color for text
   return lume <= 125 ? 'white' : 'black';
 };
